New update

+++ src/main/java/com/BaiTapLab/RestController/VoucherRestController.java
@@ -28,10 +28,14 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.BaiTapLab.Entity.DanhGia;
+import com.BaiTapLab.Entity.HanhDong;
 import com.BaiTapLab.Entity.SanPham;
 import com.BaiTapLab.Entity.Users;
 import com.BaiTapLab.Entity.Voucher;
 import com.BaiTapLab.Entity.VoucherDetail;
+import com.BaiTapLab.Repository.HanhDongRepository;
+import com.BaiTapLab.Repository.UsersRepository;
 import com.BaiTapLab.Repository.VoucherRepository;
 import com.BaiTapLab.Service.UsersService;
 import com.BaiTapLab.Service.VoucherDetailService;
@@ -53,6 +57,12 @@ public class VoucherRestController {
 	@Autowired
 	UsersService usersService;
 	
+	@Autowired
+	UsersRepository usersRepository;
+	
+	@Autowired
+	HanhDongRepository hanhDongRepository;
+	// Voucher của Thành
 	@GetMapping("/listVoucher")
     public ResponseEntity<List<List<Object>>> getVoucher() {
         List<Object[]> results = voucherRepository.findAllVouchersWithDetails();
@@ -68,6 +78,67 @@ public class VoucherRestController {
 
         return ResponseEntity.ok(listVouchers);
     }
+	
+	@GetMapping("/voucher/list/nhatky")
+	public ResponseEntity<List<Map<String, Object>>> getListNhatKy() {
+	    List<Object[]> listDanhgia = voucherRepository.listNhatKy();
+	    List<Map<String, Object>> result = new ArrayList<>();
+	    
+	    for (Object[] obj : listDanhgia) {
+	        Map<String, Object> map = new HashMap<>();
+	        map.put("ten_hanh_dong", obj[0]);
+	        map.put("ngay_hanh_dong", obj[1]);
+	        map.put("voucherID", obj[2]);
+	        map.put("accountID", obj[3]);
+	        result.add(map);
+	    }
+
+	    return ResponseEntity.ok(result);
+	}
+	
+	@GetMapping("/voucher/list/hethan")
+	public ResponseEntity<List<Map<String, Object>>> getListHetHan() {
+	    List<Object[]> listVoucher = voucherRepository.listVoucherHetHan();
+	    List<Map<String, Object>> result = new ArrayList<>();
+	    for (Object[] obj : listVoucher) {
+	        Map<String, Object> map = new HashMap<>();
+	        map.put("voucherID", obj[0]);
+	        map.put("ma_voucher", obj[1]);
+	        map.put("dieu_kien", obj[2]);
+	        map.put("don_hang_toi_thieu", obj[3]);
+	        map.put("han_su_dung", obj[4]);
+	        map.put("hinh_anh", obj[5]);
+	        map.put("so_luong", obj[6]);
+	        map.put("so_luot_SD", obj[7]);
+	        map.put("so_tien_giam", obj[8]);
+	        result.add(map);
+	    }
+	    return ResponseEntity.ok(result);
+	}
+
+	@GetMapping("/voucher/listAll")
+	public ResponseEntity<List<Map<String, Object>>> getListAll(){
+		List<Object[]> listVoucher = voucherRepository.listVoucherOnOff();
+		List<Map<String, Object>> result = new ArrayList<>();
+		for (Object[] obj : listVoucher) {
+            Map<String, Object> map = new HashMap<>();
+            map.put("voucherID", obj[0]);
+            map.put("ma_voucher", obj[1]);
+            map.put("dieu_kien", obj[2]);
+            map.put("don_hang_toi_thieu", obj[3]);
+            map.put("ngay_tao", obj[4]);
+            map.put("han_su_dung", obj[5]);
+            map.put("hinh_anh", obj[6]);
+            map.put("so_luong", obj[7]);
+            map.put("so_luot_SD", obj[8]);
+            map.put("so_tien_giam", obj[9]);
+            map.put("hoat_dong", obj[10]);
+            map.put("accountID", obj[11]);
+            map.put("trang_thai_xoa", obj[12]);
+            result.add(map);
+        }
+		return ResponseEntity.ok(result);
+	}
 	
 	@GetMapping("/voucher/getNewVoucherID")
 	public ResponseEntity<String> getnewVoucherId(){
@@ -99,66 +170,7 @@ public class VoucherRestController {
 		Optional<Voucher> voucher = voucherRepository.findById(voucherID);
 	    return voucher.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
 	}
-	
-//	@PostMapping("/voucher/add")
-//	public ResponseEntity<?> addVoucher(
-//	        @RequestParam("voucherID") String voucherID,
-//	        @RequestParam("dieu_kien") String dieu_kien,
-//	        @RequestParam("don_hang_toi_thieu") int don_hang_toi_thieu,
-//	        @RequestParam("han_su_dung") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate han_su_dung,
-//	        @RequestParam("hoat_dong") String hoat_dong,
-//	        @RequestParam("so_luong") int so_luong,
-//	        @RequestParam("so_luot_SD") int so_luot_SD,
-//	        @RequestParam("so_tien_giam") int so_tien_giam,
-//	        @RequestParam(value = "hinh_anh", required = false) MultipartFile[] hinh_anh) {
-//
-//	    try {
-//	        // Tạo đối tượng Voucher từ các tham số
-//	        Voucher voucher = new Voucher();
-//	        voucher.setVoucherID(voucherID);
-//	        voucher.setDieu_kien(dieu_kien);
-//	        voucher.setDon_hang_toi_thieu(don_hang_toi_thieu);
-//	        voucher.setHan_su_dung(han_su_dung);
-//	        voucher.setHoat_dong(hoat_dong);
-//	        voucher.setSo_luong(so_luong);
-//	        voucher.setSo_luot_SD(so_luot_SD);
-//	        voucher.setSo_tien_giam(so_tien_giam);
-//
-//	        // Xử lý file ảnh nếu được upload
-//	        if (hinh_anh != null && hinh_anh.length > 0) {
-//	            String tenHinhAnh = hinh_anh[0].getOriginalFilename();
-//	            String uploadDir = System.getProperty("user.dir") + "/uploads/images/";
-//
-//	            // Tạo thư mục nếu chưa tồn tại
-//	            File hinhFile = new File(uploadDir + tenHinhAnh);
-//	            if (!hinhFile.getParentFile().exists()) {
-//	                hinhFile.getParentFile().mkdirs();
-//	            }
-//
-//	            // Lưu file ảnh vào thư mục
-//	            hinh_anh[0].transferTo(hinhFile);
-//
-//	            // Tạo URL để truy cập ảnh và lưu vào đối tượng Voucher
-//	            String imageUrl = "http://localhost:8080/images/" + tenHinhAnh;
-//	            voucher.setHinh_anh(tenHinhAnh);
-//	        }
-//
-//	        // Lưu voucher vào DB qua service
-//	        Voucher savedVoucher = voucherService.createVoucher(voucher);
-//
-//	        // Trả về thông tin voucher đã lưu
-//	        return ResponseEntity.ok(savedVoucher);
-//
-//	    } catch (IOException e) {
-//	        // Xử lý lỗi IO
-//	        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-//	                .body("Lỗi khi lưu file: " + e.getMessage());
-//	    } catch (Exception e) {
-//	        // Xử lý lỗi chung
-//	        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-//	                .body("Lỗi khi tạo voucher: " + e.getMessage());
-//	    }
-//	}
+
 	@PostMapping("/voucher/add")
 	public ResponseEntity<?> addVoucher(
 	        @RequestParam("voucherID") String voucherID,
@@ -168,7 +180,6 @@ public class VoucherRestController {
 	        @RequestParam("ngay_tao") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate ngay_tao,
 	        @RequestParam("han_su_dung") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate han_su_dung,
 	        @RequestParam("hoat_dong") String hoat_dong,
-	        @RequestParam("hanh_dong") String hanh_dong,
 	        @RequestParam("so_luong") int so_luong,
 	        //@RequestParam("so_luot_SD") int so_luot_SD,
 	        @RequestParam("so_tien_giam") int so_tien_giam,
@@ -186,15 +197,10 @@ public class VoucherRestController {
 	        voucher.setNgay_tao(ngay_tao);
 	        voucher.setHan_su_dung(han_su_dung);
 	        voucher.setHoat_dong(hoat_dong);
-	        voucher.setHanh_dong("Thêm");
 	        voucher.setSo_luong(so_luong);
 	        //voucher.setSo_luot_SD(so_luot_SD);
 	        voucher.setSo_tien_giam(so_tien_giam);
 
-	        if (voucherService.existsByVoucherID(voucherID)) {
-	            response.put("message", "Voucher đã tồn tại, không thể thêm mới!"); // Thông báo lỗi
-	            return ResponseEntity.status(HttpStatus.CONFLICT).body(response); // 409 Conflict
-	        }
 	        // Xử lý file ảnh nếu được upload
 	        if (hinh_anh != null && hinh_anh.length > 0) {
 	            String tenHinhAnh = hinh_anh[0].getOriginalFilename();
@@ -219,8 +225,17 @@ public class VoucherRestController {
 
 	        // Tạo đối tượng VoucherDetail
 	        VoucherDetail voucherDetail = new VoucherDetail();
-	        voucherDetail.setVoucher(savedVoucher); // Liên kết VoucherDetail với Voucher vừa tạo
-
+	        voucherDetail.setVoucher(savedVoucher);
+	        
+	        voucher = voucherRepository.findById(voucherID)
+                    .orElseThrow(() -> new RuntimeException("Voucher không tồn tại"));
+	        HanhDong hanhdong = new HanhDong();
+	        hanhdong.setTen_hanh_dong("Thêm");
+	        hanhdong.setNgay_hanh_dong(LocalDate.now());
+	        hanhdong.setVoucher(voucher);
+	        hanhdong.setUsers(usersRepository.findById(accountID)
+            		.orElseThrow(() -> new RuntimeException("Account không tồn tại")));
+	        hanhDongRepository.save(hanhdong);
 	        // Tìm người dùng theo accountID (Giả sử bạn có một phương thức để tìm User)
 	        Users user = usersService.findByAccountID(accountID);
 	        if (user != null) {
@@ -257,11 +272,11 @@ public class VoucherRestController {
 	        @RequestParam("ngay_tao") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate ngay_tao,
 	        @RequestParam("han_su_dung") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate han_su_dung,
 	        @RequestParam("hoat_dong") String hoat_dong,
-	        @RequestParam("hanh_dong") String hanh_dong,
 	        @RequestParam("so_luong") int so_luong,
 	        //@RequestParam("so_luot_SD") int so_luot_SD,
 	        @RequestParam("so_tien_giam") int so_tien_giam,
-	        @RequestParam(value = "hinh_anh", required = false) MultipartFile[] hinh_anh) {
+	        @RequestParam(value = "hinh_anh", required = false) MultipartFile[] hinh_anh,
+	        @RequestParam("accountID") String accountID) {
 
 	    try {
 	        // Tìm voucher theo ID
@@ -275,7 +290,6 @@ public class VoucherRestController {
 	        voucher.setHoat_dong(hoat_dong);
 	        voucher.setNgay_tao(ngay_tao);
 	        voucher.setSo_luong(so_luong);
-	        voucher.setHanh_dong("Cập nhật");
 	        //voucher.setSo_luot_SD(so_luot_SD);
 	        voucher.setSo_tien_giam(so_tien_giam);
 
@@ -300,55 +314,73 @@ public class VoucherRestController {
 
 	        // Lưu voucher vào DB qua service
 	        Voucher updatedVoucher = voucherService.updateVoucher(voucher);
-
-	        // Trả về thông tin voucher đã lưu
+	        voucher = voucherRepository.findById(voucherID)
+                    .orElseThrow(() -> new RuntimeException("Voucher không tồn tại"));
+	        HanhDong hanhdong = new HanhDong();
+	        hanhdong.setTen_hanh_dong("Cập nhật");
+	        hanhdong.setNgay_hanh_dong(LocalDate.now());
+	        hanhdong.setVoucher(voucher);
+	        hanhdong.setUsers(usersRepository.findById(accountID)
+            		.orElseThrow(() -> new RuntimeException("Account không tồn tại")));
+	        hanhDongRepository.save(hanhdong);
 	        return ResponseEntity.ok(updatedVoucher);
 
 	    } catch (IOException e) {
-	        // Xử lý lỗi IO
 	        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
 	                .body("Lỗi khi lưu file: " + e.getMessage());
 	    } catch (Exception e) {
-	        // Xử lý lỗi chung
 	        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
 	                .body("Lỗi khi cập nhật voucher: " + e.getMessage());
 	    }
-	}
-
-	@DeleteMapping("/voucher/delete/{voucherID}")
-	public ResponseEntity<?> deleteVoucher(@PathVariable String voucherID) {
-	    try {
-	        // Tìm voucher theo ID
-	        Voucher voucher = voucherService.findByVoucherID(voucherID);
-	        if (voucher != null) {
-	            // Xóa voucher
-	            voucherService.deleteVoucher(voucherID);
-	            return ResponseEntity.ok("Voucher đã được xóa thành công!");
-	        } else {
-	            return ResponseEntity.status(HttpStatus.NOT_FOUND)
-	                    .body("Không tìm thấy voucher với ID: " + voucherID);
-	        }
-	    } catch (Exception e) {
-	        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-	                .body("Lỗi khi xóa voucher: " + e.getMessage());
-	    }
-	}
+	}	
 	
 	@PutMapping("/voucher/deleteToGarbage/{voucherID}")
-	public ResponseEntity<Object> deleteVoucherToGarbage(@PathVariable String voucherID) {
+	public ResponseEntity<Object> deleteVoucherToGarbage(
+	    @PathVariable String voucherID,
+	    @RequestParam("accountID") String accountID) {
+		System.out.println("Account nè qq: "+ accountID);
+	    // Kiểm tra accountID
+	    if (accountID == null || accountID.isEmpty()) {
+	        return ResponseEntity.badRequest()
+	            .body(Collections.singletonMap("error", "accountID is required"));
+	    }
+
+	    // Thực hiện logic xóa
 	    boolean isDeleted = voucherService.deleteVoucherById(voucherID);
+	    Voucher voucher = voucherRepository.findById(voucherID)
+	        .orElseThrow(() -> new RuntimeException("Voucher không tồn tại"));
+
+	    HanhDong hanhdong = new HanhDong();
+	    hanhdong.setTen_hanh_dong("Chuyển vào thùng rác");
+	    hanhdong.setNgay_hanh_dong(LocalDate.now());
+	    hanhdong.setVoucher(voucher);
+	    hanhdong.setUsers(usersRepository.findById(accountID)
+	        .orElseThrow(() -> new RuntimeException("Account không tồn tại")));
+
+	    hanhDongRepository.save(hanhdong);
+
 	    if (isDeleted) {
-	        // Trả về một đối tượng JSON
 	        return ResponseEntity.ok(Collections.singletonMap("message", "Voucher đã được cập nhật trạng thái xóa."));
 	    } else {
-	        // Trả về một đối tượng JSON chứa thông báo lỗi
-	        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(Collections.singletonMap("error", "Không tìm thấy voucher với ID: " + voucherID));
+	        return ResponseEntity.status(HttpStatus.NOT_FOUND)
+	            .body(Collections.singletonMap("error", "Không tìm thấy voucher với ID: " + voucherID));
 	    }
 	}
 
+
 	@PutMapping("/voucher/reloadFromGarbage/{voucherID}")
-	public ResponseEntity<Object> deleteVoucherToGarbageNull(@PathVariable String voucherID) {
+	public ResponseEntity<Object> deleteVoucherToGarbageNull(@PathVariable String voucherID,
+			@RequestParam("accountID") String accountID) {
 	    boolean isDeleted = voucherService.reloadVoucherById(voucherID);
+	    Voucher voucher = voucherRepository.findById(voucherID)
+		        .orElseThrow(() -> new RuntimeException("Voucher không tồn tại"));
+	    HanhDong hanhdong = new HanhDong();
+	    hanhdong.setTen_hanh_dong("Khôi phục voucher");
+	    hanhdong.setNgay_hanh_dong(LocalDate.now());
+	    hanhdong.setVoucher(voucher);
+	    hanhdong.setUsers(usersRepository.findById(accountID)
+		        .orElseThrow(() -> new RuntimeException("Account không tồn tại")));
+	    hanhDongRepository.save(hanhdong);
 	    if (isDeleted) {
 	        // Trả về một đối tượng JSON
 	        return ResponseEntity.ok(Collections.singletonMap("message", "Voucher đã được cập nhật trạng thái xóa."));
@@ -368,7 +400,6 @@ public class VoucherRestController {
 			Voucher voucher = voucherService.findByVoucherID(voucherID);
 			if (voucher != null) {
 			// Cập nhật hành động
-			voucher.setHanh_dong(hanhDong);
 			voucher.setNgay_tao(LocalDate.now());
 			voucherRepository.save(voucher);
 			
@@ -384,3 +415,22 @@ public class VoucherRestController {
 	}
 	
 }
+
+//@DeleteMapping("/voucher/delete/{voucherID}")
+//public ResponseEntity<?> deleteVoucher(@PathVariable String voucherID) {
+//    try {
+//        // Tìm voucher theo ID
+//        Voucher voucher = voucherService.findByVoucherID(voucherID);
+//        if (voucher != null) {
+//            // Xóa voucher
+//            voucherService.deleteVoucher(voucherID);
+//            return ResponseEntity.ok("Voucher đã được xóa thành công!");
+//        } else {
+//            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+//                    .body("Không tìm thấy voucher với ID: " + voucherID);
+//        }
+//    } catch (Exception e) {
+//        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+//                .body("Lỗi khi xóa voucher: " + e.getMessage());
+//    }
+//}
